% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{make_patterns}
\alias{make_patterns}
\title{Generate All Binary Attribute Patterns}
\usage{
make_patterns(K)
}
\arguments{
\item{K}{Integer specifying the number of attributes.}
}
\value{
A binary matrix of dimension \eqn{2^K \times K}, where each row
  represents a unique attribute pattern.
}
\description{
Generates all possible binary attribute mastery patterns of length \eqn{K}.
Each pattern represents a unique combination of mastered (1) and
non-mastered (0) attributes.
}
\details{
The function enumerates all \eqn{2^K} binary vectors of length \eqn{K},
corresponding to the full latent class space in cognitive diagnosis models.
Patterns are constructed by iterating over all possible numbers of mastered
attributes (from 0 to \eqn{K}) and generating combinations of attribute indices.

The resulting matrix has:
\itemize{
  \item \eqn{2^K} rows (all possible patterns)
  \item \eqn{K} columns (attributes)
}
}
\examples{
make_patterns(2)
#      [,1] [,2]
# [1,]    0    0
# [2,]    1    0
# [3,]    0    1
# [4,]    1    1

make_patterns(3)
# 8 possible patterns

}
